add RADOS to cmake build infrastructure.   since rados doesn't provide any packaging info (neither cmake config module nor autotool .pc file), we roll our own find module that makes an imported target for the library and hope that's good enough.
turn on rados with -DPDLFS_RADOS=ON
include and lib hints can be provided with RADOS_INCLUDE_DIR and
RADOS_LIBRARY_DIR if needed (it uses find_path and find_library,
so the usual variables that set the search path for those also apply).

the rados find module is in cmake/FindRados.cmake

revised the config file to indicate if rados is a component or not
for cmake projects that build from us in an installed location.