ran across packages that install differently.   for example, glog from "apt-get install libgoogle-glog-dev" installs a autoconf libglog.pc file.   but if you download and install it from github using cmake, then it installs a "glog" cmake package.
we'd like to be able to use it either way, and under the same
target name.  add a xdual_import() call to xpkg-import.cmake to
handle this.  you call it with the target name, cmake package
name, and pkg-config package name, like this:

  xdual_import (glog::glog,glog,libglog)

which says try and create target "glog::glog" using find_package(glog).
if that fails, then try and create the "glog::glog" target using xpkg
import of "libglog" ... you can omit the names if they are the same.
(xdual_import is current hardwired to REQUIRED)

add a new list of depends for xdual in the config file.